+++ test/controllers/subjects_controller_test.rb
@@ -50,15 +50,10 @@ class SubjectsControllerTest < ActionController::TestCase
     login(users(:site_one_editor))
     assert_difference('Sheet.current.where(missing: true).count', 1) do
       post :set_sheet_as_missing, id: @subject, project_id: @project, design_id: designs(:all_variable_types),
-                                  subject_event_id: subject_events(:one)
+                                  subject_event_id: subject_events(:one), format: 'js'
     end
-    assert_redirected_to event_project_subject_path(
-      assigns(:project),
-      assigns(:subject),
-      event_id: assigns(:sheet).subject_event.event,
-      subject_event_id: assigns(:sheet).subject_event,
-      event_date: assigns(:sheet).subject_event.event_date_to_param
-    )
+    assert_template 'set_sheet_as_missing'
+    assert_response :success
   end
 
   test 'should get send url as site editor' do